# openapi.yaml
openapi: 3.0.3
info:
  title: NexusID API
  version: 0.1.0
  description: OAuth2/OIDC, admin APIs, and user portal APIs.
servers:
  - url: http://127.0.0.1:4000
paths:
...
  /saml/metadata:
    get:
      summary: Generate SAML metadata for a service provider
      parameters:
        - in: query
          name: spId
          required: true
          schema:
            type: string
          description: Service provider ID
      responses:
        '200':
          description: SAML metadata XML
          content:
            application/samlmetadata+xml:
              schema:
                type: string
  /saml/sso:
    post:
      summary: Create signed SAML response for SP login handoff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spId]
              properties:
                spId:
                  type: string
                requestId:
                  type: string
                  description: Optional request correlation from upstream AuthnRequest
                relayState:
                  type: string
                userId:
                  type: string
                  description: Optional override user; defaults to current session user
                responseMode:
                  type: string
                  enum: [form_post, json]
                  default: form_post
      responses:
        '200':
          description: HTML auto-post form or JSON payload with base64 signed SAML response
        '401':
          description: No active session and no valid user context
        '400':
          description: Invalid service provider or payload
  /saml/acs/{spId}:
    post:
      summary: Verify signature, validate, and audit inbound SAML assertion response
      parameters:
        - in: path
          name: spId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [SAMLResponse]
              properties:
                SAMLResponse:
                  type: string
                  description: Base64-encoded SAML response XML
                RelayState:
                  type: string
      responses:
        '200':
          description: Assertion accepted and audited
        '400':
          description: Assertion validation failed
        '422':
          description: Assertion security validation failed (signature/signature-wrapping/audience/destination/expiry/replay)
        '404':
          description: Service provider not found
  /saml/slo:
    post:
      summary: Revoke current session via SAML single logout handoff
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
                relayState:
                  type: string
                reason:
                  type: string
      responses:
        '200':
          description: Logout processed
...
  /api/admin/service-identities:
    get:
      summary: List service identities
      tags: [Workload Identity]
      responses:
        '200':
          description: List of service identities
    post:
      summary: Create a service identity
      tags: [Workload Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [active, inactive, suspended]
                allowedScopes:
                  type: array
                  items:
                    type: string
                allowedAudiences:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Service identity created
  /api/admin/service-identities/{id}:
    get:
      summary: Get service identity with credentials
      tags: [Workload Identity]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service identity detail
    patch:
      summary: Update service identity
      tags: [Workload Identity]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated service identity
    delete:
      summary: Delete service identity
      tags: [Workload Identity]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /api/admin/service-identities/{id}/credentials:
    post:
      summary: Issue new credential for service identity
      tags: [Workload Identity]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresInDays:
                  type: integer
      responses:
        '201':
          description: Credential issued with plainClientSecret (shown once)
  /api/admin/service-identities/{id}/credentials/rotate:
    post:
      summary: Rotate credential (issue + revoke old)
      tags: [Workload Identity]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credentialId]
              properties:
                credentialId:
                  type: string
                expiresInDays:
                  type: integer
      responses:
        '201':
          description: New credential issued
  /api/admin/service-identities/{id}/credentials/{credentialId}:
    delete:
      summary: Revoke a credential
      tags: [Workload Identity]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: credentialId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Revoked
  /api/admin/service-identities/{id}/usage:
    get:
      summary: Get credential usage for a service identity
      tags: [Workload Identity]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credential usage telemetry
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT